return CustomPropertiesStream to repository_streams
