+++ tap_github/organization_streams.py
@@ -5,7 +5,6 @@ from typing import Any, Dict, Iterable, List, Optional
 from singer_sdk import typing as th  # JSON Schema typing helpers
 
 from tap_github.client import GitHubRestStream
-from tap_github.repository_streams import RepositoryStream
 
 
 class OrganizationStream(GitHubRestStream):
@@ -177,26 +176,4 @@ class TeamRolesStream(GitHubRestStream):
         th.Property("url", th.StringType),
         th.Property("role", th.StringType),
         th.Property("state", th.StringType),
-    ).to_dict()
-
-
-class CustomPropertiesStream(GitHubRestStream):
-    """Defines 'custom_properties' stream."""
-
-    name = "custom_properties"
-    path = "/repos/{org}/{repo}/properties/values"
-    primary_keys = ["repo", "org", "property_name"]
-    replication_key = None
-    parent_stream_type = RepositoryStream
-    ignore_parent_replication_key = True
-    state_partitioning_keys = ["repo", "org"]
-
-    schema = th.PropertiesList(
-        # Parent Keys
-        th.Property("repo", th.StringType),
-        th.Property("org", th.StringType),
-        th.Property("repo_id", th.IntegerType),
-        # Custom Property Keys
-        th.Property("property_name", th.StringType),
-        th.Property("value", th.StringType),
     ).to_dict()
