+++ tap_github/streams.py
@@ -4,7 +4,6 @@ from typing import List, Set, Type
 from singer_sdk.streams.core import Stream
 
 from tap_github.organization_streams import (
-    CustomPropertiesStream,
     OrganizationStream,
     TeamMembersStream,
     TeamRolesStream,
@@ -18,6 +17,7 @@ from tap_github.repository_streams import (
     CommitsStream,
     CommunityProfileStream,
     ContributorsStream,
+    CustomPropertiesStream,
     DependenciesStream,
     DependentsStream,
     EventsStream,
@@ -75,6 +75,7 @@ class Streams(Enum):
             CommitsStream,
             CommunityProfileStream,
             ContributorsStream,
+            CustomPropertiesStream,
             DependenciesStream,
             DependentsStream,
             EventsStream,
